Add files via upload
This is code underpins a classification workflow designed to be robust to changes in the spatial resolution of data. The workflow is flexible to image variables and derivatives and currently supported multi threaded processing. Any classifier included in the sklearn library can be implemented as part of this workflow. We have maintained the .ipynb to allow integration into a jupyter notebook working environment.